Refactor packet dispatcher out of snet
